requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
cryptography==41.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
